Norm ok finally
